requirements.txt
black==24.8.0
click==8.1.7
flake8==7.1.1
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.2

python-dotenv~=1.0.1
pip~=24.0
protobuf~=3.20.3
boto3~=1.34.146
pytz~=2024.1